Slightly retarded, but working versioning! added Seconds to timeout, updadted system message, bugfixes... and TIMEOUTS!!!!

+++ GladosV3.Module.ImageGenerator/ImageGeneratorModule.cs
@@ -18,16 +18,24 @@ namespace GladosV3.Module.ImageGeneration
         [Command("delete", RunMode = RunMode.Async)]
         [Remarks("delete")]
         [Summary("delete")]
+        [Attributes.Timeout(1, 30, Attributes.Measure.Seconds)]
         public async Task Delete([Remainder]string text)
         {
-            await Context.Channel.SendFileAsync(_service.Delete(text,Context).GetAwaiter().GetResult(),"delet.jpg");
+            if (!_service.fail)
+                await Context.Channel.SendFileAsync(_service.Delete(text, Context).GetAwaiter().GetResult(), "delet.jpg");
+            else
+                await Context.Channel.SendMessageAsync("There was an error... Check the logs!");
         }
         [Command("shit", RunMode = RunMode.Async)]
         [Remarks("shit")]
         [Summary("shit")]
+        [Attributes.Timeout(1,30, Attributes.Measure.Seconds)]
         public async Task Shit([Remainder]string text)
         {
-            await Context.Channel.SendFileAsync(_service.Shit(text.Split(','),Context).GetAwaiter().GetResult(), "shit.jpg");
+            if (!_service.fail)
+                await Context.Channel.SendFileAsync(_service.Shit(text.Split(','), Context).GetAwaiter().GetResult(), "shit.jpg");
+            else
+                await Context.Channel.SendMessageAsync("There was an error... Check the logs!");
         }
         
     }
